Avoid shell for encryption command line.
Because the cleartext and the password were passed on the command line, others
on the same system could see them in ps(1) output, top(1), et cetera.

The password is now passed in an environment variable. At least for Linux, this
fixes the problem, but there may be operating systems where users can see the
environment variables of others. A further improvement could be to pass a file
descriptor instead.

Also, the approach with open2 instead of qx`` eliminates the need for a shell
to parse the command line, a minor performance improvement.

This commit also removes the unused and broken "decrypt" subroutine, and the
no longer necessary "sanitize" sub, also eliminating the scary regexes :)