Add tmux integration, sharing setting with screen integration.
Adds support to irssinotifier.pl script to detect if tmux is attached and
actually if the pane containing irssi is currently active.  The setting
controlling this is renamed from screen_detached_only to mplexer_detached_only
and is used for both multiplexer integration features.  The feature only
works if the mplexer is behaving normally in terms of setting the expected
environment variables; otherwise, the setting has not effect and the
script behaves as though irssi is not being run under a persistent terminal
multiplexer.
Also, this resolves issue #155 as there is no attempt to run either screen
or tmux executables unless the expected env variables are set.

Note that the tmux support depends on tmux version 1.8 or greater.